refactor: add import aliases and remove some cycles
