+++ .knip.json
@@ -5,12 +5,12 @@
       "ignoreDependencies": ["typescript-eslint", "supports-color"]
     },
     "webapp": {
-      "ignore": ["env.d.ts"],
       "ignoreDependencies": ["buffer"]
     },
     "isomorphic-wrtc": {},
     "cli": {
       "entry": [
+        "src/cli.ts",
         "src/benchmark_gpt.ts",
         "src/hellaswag_gpt.ts",
         "src/train_gpt.ts"
