+++ discojs-node/tsconfig.vitest.json
@@ -2,7 +2,7 @@
   "extends": "../tsconfig.base.json",
   "references": [
     {
-      "path": "../discojs"
+      "path": "../discojs/tsconfig.vitest.json"
     }
   ],
   "compilerOptions": { "noEmit": true },
