+++ discojs-web/package.json
@@ -2,11 +2,15 @@
   "name": "@epfml/discojs-web",
   "version": "3.0.0",
   "type": "module",
-  "exports": "./dist/index.js",
+  "exports": {
+    "@disco/source": "./src/index.ts",
+    "default": "./dist/index.js",
+    "types": "./dist/index.d.ts"
+  },
   "types": "dist/index.d.ts",
   "scripts": {
     "watch": "nodemon --ext ts --ignore dist --watch ../discojs/dist --watch . --exec pnpm run",
-    "build": "tsc --build",
+    "build": "tsc --build tsconfig.lib.json",
     "test": "cd .. && vitest --run --project=discojs-web"
   },
   "repository": {
