+++ discojs-web/tsconfig.vitest.json
@@ -2,12 +2,13 @@
   "extends": "../tsconfig.base.json",
   "references": [
     {
-      "path": "../discojs"
+      "path": "../discojs/tsconfig.vitest.json"
     }
   ],
   "compilerOptions": {
     "lib": ["DOM"],
-    "noEmit": true
+    "noEmit": true,
+    "composite": true
   },
   "include": ["src"]
 }
