+++ discojs/package.json
@@ -2,11 +2,17 @@
   "name": "@epfml/discojs",
   "version": "3.0.0",
   "type": "module",
-  "exports": "./dist/index.js",
+  "exports": {
+    ".": {
+      "@disco/source": "./src/index.ts",
+      "default": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    }
+  },
   "types": "dist/index.d.ts",
   "scripts": {
     "watch": "nodemon --ext ts --ignore dist --exec pnpm run",
-    "build": "tsc --build",
+    "build": "tsc --build tsconfig.lib.json",
     "test": "cd .. && vitest --run --project=discojs"
   },
   "repository": {
@@ -15,6 +21,33 @@
   },
   "bugs": {
     "url": "https://github.com/epfml/disco/issues"
+  },
+  "imports": {
+    "#models/*": {
+      "@disco/source": "./src/models/*.ts",
+      "default": "./dist/models/*.js",
+      "types": "./dist/models/*.d.ts"
+    },
+    "#task/*": {
+      "@disco/source": "./src/task/*.ts",
+      "default": "./dist/task/*.js",
+      "types": "./dist/task/*.d.ts"
+    },
+    "#weights/*": {
+      "@disco/source": "./src/weights/*.ts",
+      "default": "./dist/weights/*.js",
+      "types": "./dist/weights/*.d.ts"
+    },
+    "#dtypes/*": {
+      "@disco/source": "./src/types/*.ts",
+      "default": "./dist/types/*.js",
+      "types": "./dist/types/*.d.ts"
+    },
+    "#dataset/*": {
+      "@disco/source": "./src/dataset/*.ts",
+      "default": "./dist/dataset/*.js",
+      "types": "./dist/dataset/*.d.ts"
+    }
   },
   "homepage": "https://github.com/epfml/disco#readme",
   "dependencies": {
