+++ discojs/src/default_tasks/mnist.ts
@@ -1,7 +1,5 @@
-import * as tf from "@tensorflow/tfjs";
-
-import type { Model, TaskProvider } from "../index.js";
-import { models } from "../index.js";
+import type { TaskProvider } from "#task/index";
+import { cards } from "#models/index";
 
 export const mnist: TaskProvider<"image", "decentralized"> = {
   getTask() {
@@ -82,8 +80,8 @@ export const mnist: TaskProvider<"image", "decentralized"> = {
       metrics: ["accuracy"],
     });
 
-    return Promise.resolve(new models.TFJS("image", model));
+    return Promise.resolve(new TFJS("image", model));
   },
 
-  modelCard: models.cards.PytorchMNISTClassifier,
+  modelCard: cards.PytorchMNISTClassifier,
 };
