+++ discojs/src/default_tasks/tinder_dog.ts
@@ -1,7 +1,5 @@
-import * as tf from "@tensorflow/tfjs";
-
-import type { Model, TaskProvider } from "../index.js";
-import { models } from "../index.js";
+import type { TaskProvider } from "#task/index";
+import { cards } from "#models/index";
 
 export const tinderDog: TaskProvider<"image", "federated"> = {
   getTask() {
@@ -95,8 +93,8 @@ export const tinderDog: TaskProvider<"image", "federated"> = {
       metrics: ["accuracy"],
     });
 
-    return Promise.resolve(new models.TFJS("image", model));
+    return Promise.resolve(new TFJS("image", model));
   },
 
-  modelCard: models.cards.DogClassifier,
+  modelCard: cards.DogClassifier,
 };
