+++ discojs/src/models/cards/LUSClassifier.ts
@@ -1,5 +1,6 @@
-import { Model, ModelCard } from "../index.js";
-import { model } from "../implementations/LUSImageClassifier.js";
+import { Model } from "#models/model";
+import { ModelCard } from "#models/model_card";
+import { model } from "#models/implementations/LUSImageClassifier";
 
 export const LUSClassifier: ModelCard<"image"> = {
   card: {
