+++ discojs/src/models/cards/mobileNetV1_cifar10.ts
@@ -1,5 +1,5 @@
-import { Model, ModelCard } from "../index.js";
-import { getModel } from "../implementations/mobileNetV1_cifar10.js";
+import { Model, ModelCard } from "#models/index";
+import { getModel } from "#models/implementations/mobileNetV1_cifar10";
 
 export const CIFAR10Classifier: ModelCard<"image"> = {
   card: {
