+++ discojs/src/models/index.ts
@@ -1,17 +1,20 @@
 export { Model } from "./model.js";
-export { BatchLogs, EpochLogs, ValidationMetrics } from "./logs.js";
+export type { BatchLogs, ValidationMetrics } from "./logs.js";
+export { EpochLogs } from "./logs.js";
 export { Tokenizer } from "./tokenizer.js";
 
 export { GPT } from "./implementations/gpt/index.js";
 export { ONNXModel } from "./onnx.js";
-export { GPTConfig } from "./implementations/gpt/config.js";
-export {
-  evaluate as evaluate_hellaswag,
+export type { GPTConfig } from "./implementations/gpt/config.js";
+export type {
   HellaSwagDataset,
   HellaSwagExample,
+} from "./implementations/hellaswag.js";
+export {
+  evaluate as evaluate_hellaswag,
   HELLASWAG_URL,
 } from "./implementations/hellaswag.js";
 export { TFJS } from "./tfjs.js";
-export { ModelCard } from "./model_card.js";
+export type { ModelCard } from "./model_card.js";
 
 export * as cards from "./cards/index.js";
