+++ package.json
@@ -18,7 +18,7 @@
     "eslint": "10.0.3",
     "eslint-plugin-cypress": "6.2.0",
     "eslint-plugin-vue": "10.9.2",
-    "knip": "6.17.1",
+    "knip": "6.31.0",
     "prettier": "3.6.2",
     "supports-color": "8.1.1",
     "typescript": "catalog:",
