+++ server/package.json
@@ -3,11 +3,15 @@
   "private": true,
   "type": "module",
   "main": "dist/index.js",
-  "exports": "./dist/index.js",
+  "exports": {
+    "@disco/source": "./src/index.ts",
+    "default": "./dist/index.js",
+    "types": "./dist/index.d.ts"
+  },
   "scripts": {
     "watch": "nodemon --ext ts --ignore dist --watch ../discojs-node/dist --watch . --exec pnpm run",
     "start": "node dist/main.js",
-    "build": "tsc --build",
+    "build": "tsc --build tsconfig.lib.json",
     "test": "cd .. && vitest --run --project=server"
   },
   "author": "",
