+++ server/tsconfig.lib.json
@@ -1,12 +1,16 @@
 {
-  "extends": "../tsconfig.base.json",
-  "compilerOptions": { "rootDir": "./src", "outDir": "dist" },
+  "extends": "../tsconfig.base.lib.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "dist",
+    "composite": true
+  },
   "references": [
     {
-      "path": "../discojs"
+      "path": "../discojs/tsconfig.lib.json"
     },
     {
-      "path": "../discojs-node"
+      "path": "../discojs-node/tsconfig.lib.json"
     }
   ],
   "include": ["src"]
