+++ server/tsconfig.vitest.json
@@ -1,13 +1,5 @@
 {
   "extends": "../tsconfig.base.json",
   "compilerOptions": { "noEmit": true },
-  "references": [
-    {
-      "path": "../discojs"
-    },
-    {
-      "path": "../discojs-node"
-    }
-  ],
   "include": ["src", "tests"]
 }
