+++ tsconfig.base.json
@@ -5,6 +5,7 @@
     "declarationMap": true,
 
     "module": "node16",
+    "moduleResolution": "node16",
     "target": "es2022",
 
     "strict": true,
