+++ tsconfig.base.lib.json
@@ -1,0 +1,6 @@
+{
+  "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "customConditions": ["@disco/source"]
+  }
+}
