+++ tsconfig.eslint.json
@@ -1,4 +1,4 @@
 {
-  "extends": "./tsconfig.base.json",
+  "extends": "./tsconfig.base.lib.json",
   "include": ["**/src", "**/tests", "**/*config.ts"]
 }
