+++ webapp/tsconfig.vitest.json
@@ -1,18 +1,18 @@
 {
   "extends": "./tsconfig.app.json",
+  // "include": ["src/**/__tests__/*", "env.d.ts"],
   "exclude": [],
   "references": [
     {
-      "path": "../discojs"
+      "path": "../discojs/tsconfig.lib.json"
     },
     {
-      "path": "../discojs-web"
+      "path": "../discojs-web/tsconfig.lib.json"
     }
   ],
   "compilerOptions": {
     "composite": true,
     "tsBuildInfoFile": "tsconfig.vitest.tsbuildinfo",
-
     "lib": [],
     "types": ["node", "jsdom"]
   }
